refactor(sync): extract per-change push CRUD ops (tech-debt)
Continues the sync_service.py split. No behavior change.

- `_sync_push_ops.py`: the per-change create/update/delete operations for
  configs and rows (push_create, push_update, push_row_change + the row
  create/update/delete helpers) as free functions taking the service. push()
  calls push_create(self, ...) / push_update(self, ...) / push_row_change(self,
  ...). The 2 direct test calls to _push_update_row were updated to the
  relocated push_update_row.

`sync_service.py`: 2244 lines. Cumulative over this PR: 4033 -> 2244 (-1789,
~44%), across 8 cohesive modules (_sync_models, _sync_bindings, _sync_clone,
_sync_writeback, _sync_storage, _sync_bulk, _sync_branch, _sync_push_ops). The
remainder is the core orchestration (init/pull/diff/push/status) + shared
branch-path/manifest helpers. make check green (4092 passed).
