+++ src/keboola_agent_cli/services/sync_service.py
@@ -4,7 +4,6 @@ Handles downloading Keboola project configurations to the local filesystem
 in a dev-friendly format (YAML configs), and tracking local changes.
 """
 
-import copy
 import hashlib
 import json
 import logging
@@ -73,6 +72,7 @@ from ._sync_bulk import (
 )
 from ._sync_clone import clone_project as _clone_project_impl
 from ._sync_models import CreatedConfig, LocalConfigHashes
+from ._sync_push_ops import push_create, push_row_change, push_update
 from ._sync_storage import (
     fetch_jobs_per_config,
     fetch_samples,
@@ -81,9 +81,7 @@ from ._sync_storage import (
 )
 from ._sync_writeback import (
     propagate_kbc_metadata,
-    writeback_after_push,
     writeback_create_config_in_manifest,
-    writeback_create_row_in_manifest,
 )
 from .base import BaseService
 
@@ -1485,7 +1483,8 @@ class SyncService(BaseService):
 
                 try:
                     if change_type == "added":
-                        result = self._push_create(
+                        result = push_create(
+                            self,
                             client,
                             component_id,
                             config_path_str,
@@ -1540,7 +1539,8 @@ class SyncService(BaseService):
                             pushed_details.append(change)
 
                     elif change_type == "modified":
-                        self._push_update(
+                        push_update(
+                            self,
                             client,
                             component_id,
                             config_id,
@@ -1613,7 +1613,8 @@ class SyncService(BaseService):
                 )
 
                 try:
-                    new_row_id = self._push_row_change(
+                    new_row_id = push_row_change(
+                        self,
                         client,
                         change_type=change_type,
                         component_id=component_id,
@@ -1781,344 +1782,6 @@ class SyncService(BaseService):
                 "message": str(exc),
             }
         )
-
-    def _push_row_change(
-        self,
-        client: Any,
-        *,
-        change_type: str,
-        component_id: str,
-        parent_config_id: str,
-        row_id: str,
-        row_path_str: str,
-        project_root: Path,
-        manifest: Manifest,
-        branch_id: int | None,
-        allow_plaintext_fallback: bool = False,
-    ) -> str | None:
-        """Dispatch a single row-level change (added/modified/deleted) to the API.
-
-        ``#``-prefixed secrets in the row's configuration are encrypted via
-        :func:`encrypt_secrets_in_config` before POST/PUT (same fail-closed
-        semantics as parent configs). Mutates ``manifest`` in place; the
-        caller is responsible for persisting it.
-
-        ``parent_config_id`` must already be the *effective* parent id: on a
-        fresh CREATE the caller remaps the diff-time placeholder to the
-        API-assigned ULID before dispatch, so both the manifest parent lookup
-        and ``create_config_row(config_id=...)`` hit the real config (KFR-05).
-
-        Returns the API-assigned row id on ``added`` (so the caller can map
-        placeholder -> ULID for variable-link backfill), else ``None``.
-        """
-        parent = next(
-            (
-                c
-                for c in manifest.configurations
-                if c.component_id == component_id and c.id == parent_config_id
-            ),
-            None,
-        )
-        if parent is None and change_type != "deleted":
-            raise KeboolaApiError(
-                message=(
-                    f"Cannot push row {row_id}: parent config {component_id}/"
-                    f"{parent_config_id} is not tracked in the manifest."
-                ),
-                status_code=0,
-                error_code=ErrorCode.PARENT_CONFIG_NOT_TRACKED,
-            )
-
-        project_id = manifest.project.id if manifest.project else None
-
-        if change_type == "deleted":
-            self._push_delete_row(
-                client,
-                component_id=component_id,
-                parent_config_id=parent_config_id,
-                row_id=row_id,
-                parent=parent,
-                branch_id=branch_id,
-            )
-            return None
-
-        # added / modified both read a local row file and encrypt-then-push.
-        assert parent is not None  # guarded above for non-deleted change_types
-        source_branch_path = self._resolve_source_branch_path(manifest, project_root, branch_id)
-        row_dir = project_root / source_branch_path / parent.path / row_path_str
-
-        if change_type == "added":
-            return self._push_create_row(
-                client,
-                component_id=component_id,
-                parent_config_id=parent_config_id,
-                row_dir=row_dir,
-                parent=parent,
-                row_path_str=row_path_str,
-                branch_id=branch_id,
-                project_id=project_id,
-                allow_plaintext_fallback=allow_plaintext_fallback,
-            )
-
-        if change_type == "modified":
-            self._push_update_row(
-                client,
-                component_id=component_id,
-                parent_config_id=parent_config_id,
-                row_id=row_id,
-                row_dir=row_dir,
-                parent=parent,
-                branch_id=branch_id,
-                project_id=project_id,
-                allow_plaintext_fallback=allow_plaintext_fallback,
-            )
-            return None
-
-        raise ValueError(f"Unsupported row change_type: {change_type}")
-
-    def _push_create_row(
-        self,
-        client: Any,
-        *,
-        component_id: str,
-        parent_config_id: str,
-        row_dir: Path,
-        parent: ManifestConfiguration,
-        row_path_str: str,
-        branch_id: int | None,
-        project_id: int | None,
-        allow_plaintext_fallback: bool,
-    ) -> str:
-        """POST a new row; record API-assigned id + hashes in the parent's row list.
-
-        Returns the API-assigned row id.
-        """
-        local_data = self._read_config_file(row_dir)
-        if local_data is None:
-            raise FileNotFoundError(f"Row file not found: {row_dir / CONFIG_FILENAME}")
-
-        pristine_data = copy.deepcopy(local_data)
-        name, description, configuration = local_row_to_api(local_data, component_id)
-        configuration = encrypt_secrets_in_config(
-            client,
-            project_id,
-            component_id,
-            configuration,
-            allow_plaintext_fallback=allow_plaintext_fallback,
-        )
-
-        result = client.create_config_row(
-            component_id=component_id,
-            config_id=parent_config_id,
-            name=name,
-            configuration=configuration,
-            description=description,
-            branch_id=branch_id,
-        )
-        new_row_id = str(result.get("id", ""))
-        logger.info("Created row %s/%s/%s", component_id, parent_config_id, new_row_id)
-
-        # Write-back: encrypted secrets land in the local file so a subsequent
-        # diff sees local == remote. ``config_id=""`` tells the shared helper
-        # to skip writing a config_id into ``_keboola`` (rows use ``row_id``).
-        writeback_after_push(self, pristine_data, row_dir, "", configuration)
-
-        row_file = row_dir / CONFIG_FILENAME
-        new_file_hash = self._file_hash(row_file) if row_file.exists() else ""
-        cfg_hash_value = config_hash(pristine_data)
-        writeback_create_row_in_manifest(
-            parent=parent,
-            row_path_str=row_path_str,
-            new_row_id=new_row_id,
-            file_hash=new_file_hash,
-            cfg_hash=cfg_hash_value,
-        )
-        return new_row_id
-
-    def _push_update_row(
-        self,
-        client: Any,
-        *,
-        component_id: str,
-        parent_config_id: str,
-        row_id: str,
-        row_dir: Path,
-        parent: ManifestConfiguration,
-        branch_id: int | None,
-        project_id: int | None,
-        allow_plaintext_fallback: bool,
-    ) -> None:
-        """PUT an existing row; refresh its hashes in the parent's row list."""
-        local_data = self._read_config_file(row_dir)
-        if local_data is None:
-            raise FileNotFoundError(f"Row file not found: {row_dir / CONFIG_FILENAME}")
-
-        pristine_data = copy.deepcopy(local_data)
-        name, description, configuration = local_row_to_api(local_data, component_id)
-        configuration = encrypt_secrets_in_config(
-            client,
-            project_id,
-            component_id,
-            configuration,
-            allow_plaintext_fallback=allow_plaintext_fallback,
-        )
-
-        client.update_config_row(
-            component_id=component_id,
-            config_id=parent_config_id,
-            row_id=row_id,
-            name=name,
-            configuration=configuration,
-            description=description,
-            change_description="Updated via kbagent sync push",
-            branch_id=branch_id,
-        )
-        logger.info("Updated row %s/%s/%s", component_id, parent_config_id, row_id)
-
-        writeback_after_push(self, pristine_data, row_dir, "", configuration)
-
-        row_file = row_dir / CONFIG_FILENAME
-        new_file_hash = self._file_hash(row_file) if row_file.exists() else ""
-        cfg_hash_value = config_hash(pristine_data)
-        for r in parent.rows:
-            if r.id == row_id:
-                r.metadata["pull_hash"] = new_file_hash
-                r.metadata["pull_config_hash"] = cfg_hash_value
-                break
-
-    def _push_delete_row(
-        self,
-        client: Any,
-        *,
-        component_id: str,
-        parent_config_id: str,
-        row_id: str,
-        parent: ManifestConfiguration | None,
-        branch_id: int | None,
-    ) -> None:
-        """DELETE a row; prune it from the parent's row list in the manifest."""
-        client.delete_config_row(
-            component_id=component_id,
-            config_id=parent_config_id,
-            row_id=row_id,
-            branch_id=branch_id,
-        )
-        if parent is not None:
-            parent.rows = [r for r in parent.rows if r.id != row_id]
-        logger.info("Deleted row %s/%s/%s", component_id, parent_config_id, row_id)
-
-    def _push_create(
-        self,
-        client: Any,
-        component_id: str,
-        config_path_str: str,
-        project_root: Path,
-        manifest: Manifest,
-        branch_id: int | None,
-        *,
-        allow_plaintext_fallback: bool = False,
-    ) -> dict[str, Any] | None:
-        """Create a new config from a local _config.yml file."""
-        branch_path = self._resolve_source_branch_path(manifest, project_root, branch_id)
-        config_dir = project_root / branch_path / config_path_str
-        local_data = self._read_config_file(config_dir)
-        if local_data is None:
-            return None
-
-        # Preserve pristine data for writeback (merge_code_files mutates
-        # local_data by injecting parameters.blocks which should not end
-        # up in _config.yml).
-        pristine_data = copy.deepcopy(local_data)
-
-        # Merge code files (transform.sql, transform.py, code.py) back into config
-        merge_code_files(component_id, local_data, config_dir)
-
-        name, description, configuration = local_config_to_api(local_data)
-
-        # Encrypt #-prefixed secrets before sending to API
-        project_id = manifest.project.id if manifest.project else None
-        configuration = encrypt_secrets_in_config(
-            client,
-            project_id,
-            component_id,
-            configuration,
-            allow_plaintext_fallback=allow_plaintext_fallback,
-        )
-
-        result = client.create_config(
-            component_id=component_id,
-            name=name,
-            configuration=configuration,
-            description=description,
-            branch_id=branch_id,
-        )
-        new_config_id = result.get("id", "")
-        logger.info(
-            "Created config %s/%s (ID: %s)",
-            component_id,
-            name,
-            new_config_id,
-        )
-
-        # Write back: update local file with config_id + encrypted secrets.
-        # Use pristine_data so blocks/code stay only in their code files.
-        writeback_after_push(self, pristine_data, config_dir, new_config_id, configuration)
-
-        return result
-
-    def _push_update(
-        self,
-        client: Any,
-        component_id: str,
-        config_id: str,
-        config_path_str: str,
-        project_root: Path,
-        manifest: Manifest,
-        branch_id: int | None,
-        *,
-        allow_plaintext_fallback: bool = False,
-    ) -> None:
-        """Update an existing config from a local _config.yml file."""
-        branch_path = self._resolve_source_branch_path(manifest, project_root, branch_id)
-        config_dir = project_root / branch_path / config_path_str
-        local_data = self._read_config_file(config_dir)
-        if local_data is None:
-            raise FileNotFoundError(f"Config file not found: {config_dir / CONFIG_FILENAME}")
-
-        # Preserve pristine data for writeback (merge_code_files mutates
-        # local_data by injecting parameters.blocks which should not end
-        # up in _config.yml).
-        pristine_data = copy.deepcopy(local_data)
-
-        # Merge code files (transform.sql, transform.py, code.py) back into config
-        merge_code_files(component_id, local_data, config_dir)
-
-        name, description, configuration = local_config_to_api(local_data)
-
-        # Encrypt #-prefixed secrets before sending to API
-        project_id = manifest.project.id if manifest.project else None
-        configuration = encrypt_secrets_in_config(
-            client,
-            project_id,
-            component_id,
-            configuration,
-            allow_plaintext_fallback=allow_plaintext_fallback,
-        )
-
-        client.update_config(
-            component_id=component_id,
-            config_id=config_id,
-            name=name,
-            configuration=configuration,
-            description=description,
-            change_description="Updated via kbagent sync push",
-            branch_id=branch_id,
-        )
-        logger.info("Updated config %s/%s", component_id, config_id)
-
-        # Write back: update local file with encrypted secrets.
-        # Use pristine_data so blocks/code stay only in their code files.
-        writeback_after_push(self, pristine_data, config_dir, config_id, configuration)
 
     # ------------------------------------------------------------------
     # bulk operations (all projects)
